requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
streamlit
pandas
python-dotenv
sentence-transformers